refactor(cmd/up): split functions across file to reduce file size
