+++ cmd/up_ssh.go
@@ -1,0 +1,358 @@
+package cmd
+
+import (
+	"context"
+	"fmt"
+	"os"
+	"os/exec"
+	"slices"
+	"strings"
+
+	"github.com/sirupsen/logrus"
+	"github.com/skevetter/devpod/pkg/agent"
+	client2 "github.com/skevetter/devpod/pkg/client"
+	"github.com/skevetter/devpod/pkg/client/clientimplementation"
+	"github.com/skevetter/devpod/pkg/config"
+	config2 "github.com/skevetter/devpod/pkg/devcontainer/config"
+	provider2 "github.com/skevetter/devpod/pkg/provider"
+	devssh "github.com/skevetter/devpod/pkg/ssh"
+	"github.com/skevetter/log"
+)
+
+type configureSSHParams struct {
+	sshConfigPath        string
+	sshConfigIncludePath string
+	user                 string
+	workdir              string
+	gpgagent             bool
+	devPodHome           string
+}
+
+func configureSSH(client client2.BaseWorkspaceClient, params configureSSHParams) error {
+	path, err := devssh.ResolveSSHConfigPath(params.sshConfigPath)
+	if err != nil {
+		return fmt.Errorf("invalid ssh config path: %w", err)
+	}
+	sshConfigPath := path
+
+	sshConfigIncludePath := params.sshConfigIncludePath
+	if sshConfigIncludePath != "" {
+		includePath, err := devssh.ResolveSSHConfigPath(sshConfigIncludePath)
+		if err != nil {
+			return fmt.Errorf("invalid ssh config include path: %w", err)
+		}
+		sshConfigIncludePath = includePath
+	}
+
+	err = devssh.ConfigureSSHConfig(devssh.SSHConfigParams{
+		SSHConfigPath:        sshConfigPath,
+		SSHConfigIncludePath: sshConfigIncludePath,
+		Context:              client.Context(),
+		Workspace:            client.Workspace(),
+		User:                 params.user,
+		Workdir:              params.workdir,
+		GPGAgent:             params.gpgagent,
+		DevPodHome:           params.devPodHome,
+		Provider:             client.Provider(),
+		Log:                  log.Default,
+	})
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func mergeDevPodUpOptions(baseOptions *provider2.CLIOptions) error {
+	oldOptions := *baseOptions
+	found, err := clientimplementation.DecodeOptionsFromEnv(
+		clientimplementation.DevPodFlagsUp,
+		baseOptions,
+	)
+	if err != nil {
+		return fmt.Errorf("decode up options: %w", err)
+	} else if found {
+		baseOptions.WorkspaceEnv = append(oldOptions.WorkspaceEnv, baseOptions.WorkspaceEnv...)
+		baseOptions.InitEnv = append(oldOptions.InitEnv, baseOptions.InitEnv...)
+		baseOptions.PrebuildRepositories = append(oldOptions.PrebuildRepositories, baseOptions.PrebuildRepositories...)
+		baseOptions.IDEOptions = append(oldOptions.IDEOptions, baseOptions.IDEOptions...)
+	}
+
+	err = clientimplementation.DecodePlatformOptionsFromEnv(&baseOptions.Platform)
+	if err != nil {
+		return fmt.Errorf("decode platform options: %w", err)
+	}
+
+	return nil
+}
+
+func mergeEnvFromFiles(baseOptions *provider2.CLIOptions) error {
+	var variables []string
+	for _, file := range baseOptions.WorkspaceEnvFile {
+		envFromFile, err := config2.ParseKeyValueFile(file)
+		if err != nil {
+			return err
+		}
+		variables = append(variables, envFromFile...)
+	}
+	baseOptions.WorkspaceEnv = append(baseOptions.WorkspaceEnv, variables...)
+
+	return nil
+}
+
+var inheritedEnvironmentVariables = []string{
+	"GIT_AUTHOR_NAME",
+	"GIT_AUTHOR_EMAIL",
+	"GIT_AUTHOR_DATE",
+	"GIT_COMMITTER_NAME",
+	"GIT_COMMITTER_EMAIL",
+	"GIT_COMMITTER_DATE",
+}
+
+func createSSHCommand(
+	ctx context.Context,
+	client client2.BaseWorkspaceClient,
+	logger log.Logger,
+	extraArgs []string,
+) (*exec.Cmd, error) {
+	execPath, err := os.Executable()
+	if err != nil {
+		return nil, err
+	}
+
+	args := []string{
+		"ssh",
+		"--user=root",
+		"--agent-forwarding=false",
+		"--start-services=false",
+		"--context",
+		client.Context(),
+		client.Workspace(),
+	}
+	if logger.GetLevel() == logrus.DebugLevel {
+		args = append(args, "--debug")
+	}
+	args = append(args, extraArgs...)
+
+	return exec.CommandContext(ctx, execPath, args...), nil
+}
+
+func setupDotfiles(
+	dotfiles, script string,
+	envFiles, envKeyValuePairs []string,
+	client client2.BaseWorkspaceClient,
+	devPodConfig *config.Config,
+	log log.Logger,
+) error {
+	dotfilesRepo := devPodConfig.ContextOption(config.ContextOptionDotfilesURL)
+	if dotfiles != "" {
+		dotfilesRepo = dotfiles
+	}
+
+	dotfilesScript := devPodConfig.ContextOption(config.ContextOptionDotfilesScript)
+	if script != "" {
+		dotfilesScript = script
+	}
+
+	if dotfilesRepo == "" {
+		log.Debug("No dotfiles repo specified, skipping")
+		return nil
+	}
+
+	log.Infof("Dotfiles git repository %s specified", dotfilesRepo)
+	log.Debug("Cloning dotfiles into the devcontainer...")
+
+	dotCmd, err := buildDotCmd(devPodConfig, dotfilesRepo, dotfilesScript, envFiles, envKeyValuePairs, client, log)
+	if err != nil {
+		return err
+	}
+	if log.GetLevel() == logrus.DebugLevel {
+		dotCmd.Args = append(dotCmd.Args, "--debug")
+	}
+
+	log.Debugf("Running dotfiles setup command: %v", dotCmd.Args)
+
+	writer := log.Writer(logrus.InfoLevel, false)
+
+	dotCmd.Stdout = writer
+	dotCmd.Stderr = writer
+
+	err = dotCmd.Run()
+	if err != nil {
+		return err
+	}
+
+	log.Infof("Done setting up dotfiles into the devcontainer")
+
+	return nil
+}
+
+func buildDotCmdAgentArguments(devPodConfig *config.Config, dotfilesRepo, dotfilesScript string, log log.Logger) []string {
+	agentArguments := []string{
+		"agent",
+		"workspace",
+		"install-dotfiles",
+		"--repository",
+		dotfilesRepo,
+	}
+
+	if devPodConfig.ContextOption(config.ContextOptionSSHStrictHostKeyChecking) == "true" {
+		agentArguments = append(agentArguments, "--strict-host-key-checking")
+	}
+
+	if log.GetLevel() == logrus.DebugLevel {
+		agentArguments = append(agentArguments, "--debug")
+	}
+
+	if dotfilesScript != "" {
+		log.Infof("Dotfiles script %s specified", dotfilesScript)
+		agentArguments = append(agentArguments, "--install-script", dotfilesScript)
+	}
+
+	return agentArguments
+}
+
+func buildDotCmd(devPodConfig *config.Config, dotfilesRepo, dotfilesScript string, envFiles, envKeyValuePairs []string, client client2.BaseWorkspaceClient, log log.Logger) (*exec.Cmd, error) {
+	sshCmd := []string{
+		"ssh",
+		"--agent-forwarding=true",
+		"--start-services=true",
+	}
+
+	envFilesKeyValuePairs, err := collectDotfilesScriptEnvKeyvaluePairs(envFiles)
+	if err != nil {
+		return nil, err
+	}
+
+	// Collect file-based and CLI options env variables names (aka keys) and
+	// configure ssh env var passthrough with send-env
+	allEnvKeyValuesPairs := slices.Concat(envFilesKeyValuePairs, envKeyValuePairs)
+	allEnvKeys := extractKeysFromEnvKeyValuePairs(allEnvKeyValuesPairs)
+	for _, envKey := range allEnvKeys {
+		sshCmd = append(sshCmd, "--send-env", envKey)
+	}
+
+	remoteUser, err := devssh.GetUser(client.WorkspaceConfig().ID, client.WorkspaceConfig().SSHConfigPath, client.WorkspaceConfig().SSHConfigIncludePath)
+	if err != nil {
+		remoteUser = "root"
+	}
+
+	agentArguments := buildDotCmdAgentArguments(devPodConfig, dotfilesRepo, dotfilesScript, log)
+	sshCmd = append(sshCmd,
+		"--user",
+		remoteUser,
+		"--context",
+		client.Context(),
+		client.Workspace(),
+		"--log-output=raw",
+		"--command",
+		agent.ContainerDevPodHelperLocation+" "+strings.Join(agentArguments, " "),
+	)
+	execPath, err := os.Executable()
+	if err != nil {
+		return nil, err
+	}
+
+	dotCmd := exec.Command(
+		execPath,
+		sshCmd...,
+	)
+
+	dotCmd.Env = append(dotCmd.Environ(), allEnvKeyValuesPairs...)
+	return dotCmd, nil
+}
+
+func extractKeysFromEnvKeyValuePairs(envKeyValuePairs []string) []string {
+	keys := []string{}
+	for _, env := range envKeyValuePairs {
+		keyValue := strings.SplitN(env, "=", 2)
+		if len(keyValue) == 2 {
+			keys = append(keys, keyValue[0])
+		}
+	}
+	return keys
+}
+
+func collectDotfilesScriptEnvKeyvaluePairs(envFiles []string) ([]string, error) {
+	keyValues := []string{}
+	for _, file := range envFiles {
+		envFromFile, err := config2.ParseKeyValueFile(file)
+		if err != nil {
+			return nil, err
+		}
+		keyValues = append(keyValues, envFromFile...)
+	}
+	return keyValues, nil
+}
+
+func setupGitSSHSignature(signingKey string, client client2.BaseWorkspaceClient, log log.Logger) error {
+	execPath, err := os.Executable()
+	if err != nil {
+		return err
+	}
+
+	remoteUser, err := devssh.GetUser(client.WorkspaceConfig().ID, client.WorkspaceConfig().SSHConfigPath, client.WorkspaceConfig().SSHConfigIncludePath)
+	if err != nil {
+		remoteUser = "root"
+	}
+
+	err = exec.Command(
+		execPath,
+		"ssh",
+		"--agent-forwarding=true",
+		"--start-services=true",
+		"--user",
+		remoteUser,
+		"--context",
+		client.Context(),
+		client.Workspace(),
+		"--command", fmt.Sprintf("devpod agent git-ssh-signature-helper %s", signingKey),
+	).Run()
+	if err != nil {
+		log.Error("failure in setting up git ssh signature helper")
+	}
+	return nil
+}
+
+func performGpgForwarding(
+	client client2.BaseWorkspaceClient,
+	log log.Logger,
+) error {
+	log.Debug("gpg forwarding enabled, performing immediately")
+
+	execPath, err := os.Executable()
+	if err != nil {
+		return err
+	}
+
+	remoteUser, err := devssh.GetUser(client.WorkspaceConfig().ID, client.WorkspaceConfig().SSHConfigPath, client.WorkspaceConfig().SSHConfigIncludePath)
+	if err != nil {
+		remoteUser = "root"
+	}
+
+	log.Info("forwarding gpg-agent")
+
+	// perform in background an ssh command forwarding the
+	// gpg agent, in order to have it immediately take effect
+	go func() {
+		err = exec.Command(
+			execPath,
+			"ssh",
+			"--gpg-agent-forwarding=true",
+			"--agent-forwarding=true",
+			"--start-services=true",
+			"--user",
+			remoteUser,
+			"--context",
+			client.Context(),
+			client.Workspace(),
+			"--log-output=raw",
+			"--command", "sleep infinity",
+		).Run()
+		if err != nil {
+			log.Error("failure in forwarding gpg-agent")
+		}
+	}()
+
+	return nil
+}
